fix: extract common things
