+++ exporter/otlp-grpc/lib/opentelemetry/exporter/otlp/grpc.rb
@@ -4,9 +4,6 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-require 'opentelemetry/exporter/otlp/grpc/exporter'
-require 'opentelemetry/exporter/otlp/grpc/version'
-
 # OpenTelemetry is an open source observability framework, providing a
 # general-purpose API, SDK, and related tools required for the instrumentation
 # of cloud-native software, frameworks, and libraries.
@@ -14,4 +11,14 @@ require 'opentelemetry/exporter/otlp/grpc/version'
 # The OpenTelemetry module provides global accessors for telemetry objects.
 # See the documentation for the `opentelemetry-api` gem for details.
 module OpenTelemetry
+  module Exporter
+    module OTLP
+      # GRPC contains the implementation for the OTLP over GRPC exporter
+      module GRPC
+      end
+    end
+  end
 end
+
+require 'opentelemetry/exporter/otlp/grpc/exporter'
+require 'opentelemetry/exporter/otlp/grpc/version'
