+++ exporter/otlp/test/opentelemetry/exporter/otlp/exporter_test.rb
@@ -281,7 +281,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
     it 'integrates with collector' do
       skip unless ENV['TRACING_INTEGRATION_TEST']
       WebMock.disable_net_connect!(allow: 'localhost')
-      span_data = create_span_data
+      span_data = OpenTelemetry::TestHelpers.create_span_data
       exporter = OpenTelemetry::Exporter::OTLP::Exporter.new(endpoint: 'http://localhost:4318', compression: 'gzip')
       result = exporter.export([span_data])
       _(result).must_equal(SUCCESS)
@@ -289,14 +289,14 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
 
     it 'retries on timeout' do
       stub_request(:post, 'http://localhost:4318/v1/traces').to_timeout.then.to_return(status: 200)
-      span_data = create_span_data
+      span_data = OpenTelemetry::TestHelpers.create_span_data
       result = exporter.export([span_data])
       _(result).must_equal(SUCCESS)
     end
 
     it 'returns TIMEOUT on timeout' do
       stub_request(:post, 'http://localhost:4318/v1/traces').to_return(status: 200)
-      span_data = create_span_data
+      span_data = OpenTelemetry::TestHelpers.create_span_data
       result = exporter.export([span_data], timeout: 0)
       _(result).must_equal(FAILURE)
     end
@@ -307,7 +307,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
       OpenTelemetry.logger = ::Logger.new(log_stream)
 
       stub_request(:post, 'http://localhost:4318/v1/traces').to_raise('something unexpected')
-      span_data = create_span_data
+      span_data = OpenTelemetry::TestHelpers.create_span_data
       result = exporter.export([span_data], timeout: 1)
 
       _(log_stream.string).must_match(
@@ -325,7 +325,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
       OpenTelemetry.logger = ::Logger.new(log_stream)
 
       stub_request(:post, 'http://localhost:4318/v1/traces').to_return(status: 200)
-      span_data = create_span_data
+      span_data = OpenTelemetry::TestHelpers.create_span_data
 
       Opentelemetry::Proto::Collector::Trace::V1::ExportTraceServiceRequest.stub(:encode, ->(_) { raise 'a little hell' }) do
         _(exporter.export([span_data], timeout: 1)).must_equal(FAILURE)
@@ -340,7 +340,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
 
     it 'returns TIMEOUT on timeout after retrying' do
       stub_request(:post, 'http://localhost:4318/v1/traces').to_timeout.then.to_raise('this should not be reached')
-      span_data = create_span_data
+      span_data = OpenTelemetry::TestHelpers.create_span_data
 
       @retry_count = 0
       backoff_stubbed_call = lambda do |**_args|
@@ -365,7 +365,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
     it 'returns FAILURE when encryption to receiver endpoint fails' do
       exporter = OpenTelemetry::Exporter::OTLP::Exporter.new(endpoint: 'https://localhost:4318/v1/traces')
       stub_request(:post, 'https://localhost:4318/v1/traces').to_raise(OpenSSL::SSL::SSLError.new('enigma wedged'))
-      span_data = create_span_data
+      span_data = OpenTelemetry::TestHelpers.create_span_data
       exporter.stub(:backoff?, ->(**_) { false }) do
         _(exporter.export([span_data])).must_equal(FAILURE)
       end
@@ -373,7 +373,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
 
     it 'exports a span_data' do
       stub_request(:post, 'http://localhost:4318/v1/traces').to_return(status: 200)
-      span_data = create_span_data
+      span_data = OpenTelemetry::TestHelpers.create_span_data
       result = exporter.export([span_data])
       _(result).must_equal(SUCCESS)
     end
@@ -384,7 +384,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
       OpenTelemetry.logger = ::Logger.new(log_stream)
 
       stub_request(:post, 'http://localhost:4318/v1/traces').to_return(status: 200)
-      span_data = create_span_data(total_recorded_attributes: 1, attributes: { 'a' => "\xC2".dup.force_encoding(::Encoding::ASCII_8BIT) })
+      span_data = OpenTelemetry::TestHelpers.create_span_data(total_recorded_attributes: 1, attributes: { 'a' => "\xC2".dup.force_encoding(::Encoding::ASCII_8BIT) })
 
       result = exporter.export([span_data])
 
@@ -405,7 +405,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
       details = [::Google::Protobuf::Any.pack(::Google::Protobuf::StringValue.new(value: 'you are a bad request'))]
       status = ::Google::Rpc::Status.encode(::Google::Rpc::Status.new(code: 1, message: 'bad request', details: details))
       stub_request(:post, 'http://localhost:4318/v1/traces').to_return(status: 400, body: status)
-      span_data = create_span_data
+      span_data = OpenTelemetry::TestHelpers.create_span_data
 
       result = exporter.export([span_data])
 
@@ -420,7 +420,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
 
     it 'handles Zlib gzip compression errors' do
       stub_request(:post, 'http://localhost:4318/v1/traces').to_raise(Zlib::DataError.new('data error'))
-      span_data = create_span_data
+      span_data = OpenTelemetry::TestHelpers.create_span_data
       exporter.stub(:backoff?, ->(**_) { false }) do
         _(exporter.export([span_data])).must_equal(FAILURE)
       end
@@ -442,7 +442,7 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
         { status: 200 }
       end
 
-      span_data = create_span_data
+      span_data = OpenTelemetry::TestHelpers.create_span_data
       result = exporter.export([span_data])
 
       _(result).must_equal(SUCCESS)
@@ -457,8 +457,8 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
         { status: 200 }
       end
 
-      span_data1 = create_span_data(resource: OpenTelemetry::SDK::Resources::Resource.create('k1' => 'v1'))
-      span_data2 = create_span_data(resource: OpenTelemetry::SDK::Resources::Resource.create('k2' => 'v2'))
+      span_data1 = OpenTelemetry::TestHelpers.create_span_data(resource: OpenTelemetry::SDK::Resources::Resource.create('k1' => 'v1'))
+      span_data2 = OpenTelemetry::TestHelpers.create_span_data(resource: OpenTelemetry::SDK::Resources::Resource.create('k2' => 'v2'))
 
       result = exporter.export([span_data1, span_data2])
 
@@ -483,11 +483,11 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
       end_timestamp = start_timestamp + 6
 
       OpenTelemetry.tracer_provider.add_span_processor(processor)
-      root = with_ids(trace_id, root_span_id) { tracer.start_root_span('root', kind: :internal, start_timestamp: start_timestamp) }
+      root = OpenTelemetry::TestHelpers.with_ids(trace_id, root_span_id) { tracer.start_root_span('root', kind: :internal, start_timestamp: start_timestamp) }
       root.status = OpenTelemetry::Trace::Status.ok
       root.finish(end_timestamp: end_timestamp)
       root_ctx = OpenTelemetry::Trace.context_with_span(root)
-      span = with_ids(trace_id, child_span_id) { tracer.start_span('child', with_parent: root_ctx, kind: :producer, start_timestamp: start_timestamp + 1, links: [OpenTelemetry::Trace::Link.new(root.context, 'attr' => 4)]) }
+      span = OpenTelemetry::TestHelpers.with_ids(trace_id, child_span_id) { tracer.start_span('child', with_parent: root_ctx, kind: :producer, start_timestamp: start_timestamp + 1, links: [OpenTelemetry::Trace::Link.new(root.context, 'attr' => 4)]) }
       span['b'] = true
       span['f'] = 1.1
       span['i'] = 2
@@ -495,11 +495,11 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
       span['a'] = [3, 4]
       span.status = OpenTelemetry::Trace::Status.error
       child_ctx = OpenTelemetry::Trace.context_with_span(span)
-      client = with_ids(trace_id, client_span_id) { tracer.start_span('client', with_parent: child_ctx, kind: :client, start_timestamp: start_timestamp + 2).finish(end_timestamp: end_timestamp) }
+      client = OpenTelemetry::TestHelpers.with_ids(trace_id, client_span_id) { tracer.start_span('client', with_parent: child_ctx, kind: :client, start_timestamp: start_timestamp + 2).finish(end_timestamp: end_timestamp) }
       client_ctx = OpenTelemetry::Trace.context_with_span(client)
-      with_ids(trace_id, server_span_id) { other_tracer.start_span('server', with_parent: client_ctx, kind: :server, start_timestamp: start_timestamp + 3).finish(end_timestamp: end_timestamp) }
+      OpenTelemetry::TestHelpers.with_ids(trace_id, server_span_id) { other_tracer.start_span('server', with_parent: client_ctx, kind: :server, start_timestamp: start_timestamp + 3).finish(end_timestamp: end_timestamp) }
       span.add_event('event', attributes: { 'attr' => 42 }, timestamp: start_timestamp + 4)
-      with_ids(trace_id, consumer_span_id) { tracer.start_span('consumer', with_parent: child_ctx, kind: :consumer, start_timestamp: start_timestamp + 5).finish(end_timestamp: end_timestamp) }
+      OpenTelemetry::TestHelpers.with_ids(trace_id, consumer_span_id) { tracer.start_span('consumer', with_parent: child_ctx, kind: :consumer, start_timestamp: start_timestamp + 5).finish(end_timestamp: end_timestamp) }
       span.finish(end_timestamp: end_timestamp)
       OpenTelemetry.tracer_provider.shutdown
 
@@ -635,25 +635,5 @@ describe OpenTelemetry::Exporter::OTLP::Exporter do
         req.body == Zlib.gzip(encoded_etsr)
       end
     end
-  end
-
-  def with_ids(trace_id, span_id)
-    OpenTelemetry::Trace.stub(:generate_trace_id, trace_id) do
-      OpenTelemetry::Trace.stub(:generate_span_id, span_id) do
-        yield
-      end
-    end
-  end
-
-  def create_span_data(name: '', kind: nil, status: nil, parent_span_id: OpenTelemetry::Trace::INVALID_SPAN_ID,
-                       total_recorded_attributes: 0, total_recorded_events: 0, total_recorded_links: 0, start_timestamp: OpenTelemetry::TestHelpers.exportable_timestamp,
-                       end_timestamp: OpenTelemetry::TestHelpers.exportable_timestamp, attributes: nil, links: nil, events: nil, resource: nil,
-                       instrumentation_library: OpenTelemetry::SDK::InstrumentationLibrary.new('', 'v0.0.1'),
-                       span_id: OpenTelemetry::Trace.generate_span_id, trace_id: OpenTelemetry::Trace.generate_trace_id,
-                       trace_flags: OpenTelemetry::Trace::TraceFlags::DEFAULT, tracestate: nil)
-    resource ||= OpenTelemetry::SDK::Resources::Resource.telemetry_sdk
-    OpenTelemetry::SDK::Trace::SpanData.new(name, kind, status, parent_span_id, total_recorded_attributes,
-                                            total_recorded_events, total_recorded_links, start_timestamp, end_timestamp,
-                                            attributes, links, events, resource, instrumentation_library, span_id, trace_id, trace_flags, tracestate)
   end
 end
